requirements-dev.txt
# Dependências de desenvolvimento e CI (P3)
-r requirements.txt

ruff>=0.4,<1.0        # linter + formatter (substitui flake8, isort, pyupgrade)
pip-audit>=2.7,<3.0   # auditoria de vulnerabilidades nas dependências
pytest>=8.0,<9.0
pytest-flask>=1.3,<2.0
coverage>=7.0,<8.0
